backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "httpx>=0.28.1",
    "google-genai>=0.1.0",
    "openai>=1.107.1,<2",
    "openai-agents>=0.3.3",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.37.0",
]
